ledger_gpt/requirements.txt
openai==1.12.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
tabulate==0.9.0
colorama==0.4.6
